Seperate graph code from main.py
